Implement a primality test for Problem 58
The problem is to find the side length of the number enumerated from
the center of a square starting from one so that the ratio of primes
to non-primes in the diagonals is less than 10%.

Since we don't know the number of primes that will be used, we
implement a miller-rabbin primality testing for this. The numbers 2,
7, and 61 is enough for the values of a as I bet that the number
produced before hitting the 10% mark will be less than 4 billion.
According to wikipedia, testing 2, 7 and 61 is enough for numbers less
than ~ 4 billion.